pyproject.toml
[project]
name = "clarity-semeval2026"
version = "0.1.0"
description = "SemEval 2026 CLARITY: Unmasking Political Question Evasions"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "torch>=2.0",
    "transformers>=4.36",
    "datasets>=2.14",
    "scikit-learn>=1.3",
    "pandas>=2.0",
    "pyarrow>=14.0",
    "pyyaml>=6.0",
    "tqdm>=4.65",
    "accelerate>=0.25",
    "sentencepiece>=0.1.99",
    "tiktoken>=0.5",
    "protobuf>=3.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "matplotlib>=3.7",
    "seaborn>=0.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/clarity"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
